+++ src/ui-win/anim/AnimRotateAxisView.cpp
@@ -1,5 +1,5 @@
 /*
-**  Copyright(C) 2017, StepToSky
+**  Copyright(C) 2019, StepToSky
 **
 **  Redistribution and use in source and binary forms, with or without
 **  modification, are permitted provided that the following conditions are met:
@@ -33,6 +33,7 @@
 #include "ui-win/AnimCalc.h"
 #include "resource/ResHelper.h"
 #include "presenters/Datarefs.h"
+#include "converters/ConverterAnimRotate.h"
 
 namespace ui {
 namespace win {
@@ -141,22 +142,26 @@ namespace win {
         }
     }
 
-    void AnimRotateAxisView::slotSelectionChange(void * param, NotifyInfo *) {
-        AnimRotateAxisView * view = reinterpret_cast<AnimRotateAxisView*>(param);
-        const auto selectedCount = view->mIp->GetSelNodeCount();
+    void AnimRotateAxisView::loadSelection() {
+        const auto selectedCount = mIp->GetSelNodeCount();
         if (selectedCount == 0) {
-            view->clearValues();
-            view->mData.clearLink();
+            clearValues();
+            mData.clearLink();
         }
         else if (selectedCount == 1) {
-            view->cSpnValue->SetValue(0.0f, FALSE);
-            view->mData.linkNode(view->mIp->GetSelNode(0));
+            cSpnValue->SetValue(0.0f, FALSE);
+            mData.linkNode(mIp->GetSelNode(0));
         }
         else {
-            view->clearValues();
-            view->mData.clearLink();
+            clearValues();
+            mData.clearLink();
         }
-        view->toWindow();
+        toWindow();
+    }
+
+    void AnimRotateAxisView::slotSelectionChange(void * param, NotifyInfo *) {
+        auto view = reinterpret_cast<AnimRotateAxisView*>(param);
+        view->loadSelection();
     }
 
     void AnimRotateAxisView::slotAnimationModeOff(void * param, NotifyInfo *) {
@@ -175,8 +180,6 @@ namespace win {
                                      reinterpret_cast<LPARAM>(this));
         if (res != nullptr) {
             registerCallbacks();
-            ctrl::Base win(res);
-            win.show();
         }
         return res != nullptr;
     }
@@ -184,6 +187,28 @@ namespace win {
     void AnimRotateAxisView::destroy() {
         unRegisterCallbacks();
         DestroyWindow(hwnd());
+    }
+
+    /**************************************************************************************************/
+    ///////////////////////////////////////////* Functions *////////////////////////////////////////////
+    /**************************************************************************************************/
+
+    void AnimRotateAxisView::active(const bool state) {
+        mIsActive = state;
+        if (!mIsActive) {
+            ctrl::Base(hwnd()).hide();
+            unRegisterCallbacks();
+            mData.clearLink();
+        }
+        else {
+            registerCallbacks();
+            loadSelection();
+            ctrl::Base(hwnd()).show();
+        }
+    }
+
+    bool AnimRotateAxisView::isActive() const {
+        return mIsActive;
     }
 
     /**************************************************************************************************/
@@ -221,6 +246,8 @@ namespace win {
         cBtnUpdate.setToolTip(sts::toString("Update animation keys. Use it to update your animation keys when you have changed your animation without pressing auto-key button."));
 
         switch (mData.mAxis) {
+            case MdAnimRot::LINEAR: cStcName.setText("Linear");
+                break;
             case MdAnimRot::X: cStcName.setText("X axis");
                 break;
             case MdAnimRot::Y: cStcName.setText("Y axis");
@@ -346,6 +373,11 @@ namespace win {
             cListKeys.addItem(sts::StrUtils::join(_T("#:"), i + 1, _T(" F:"), timeList[i] / tpt, _T(" V:"), mData.mKeyList[i]));
         }
         cListKeys.setCurrSelected(sCurrSelected);
+
+        if (mData.mAxis == MdAnimRot::LINEAR) {
+            const auto axisNum = ConverterAnimRotate::calculateLinearAxisNum(mData.linkedNode());
+            cStcName.setText("Linear (got "s.append(std::to_string(axisNum)).append(" axis)"));
+        }
     }
 
     void AnimRotateAxisView::setEnable() {